Split trigonometric/hyperbolic test from test_math.py
